refactor, moved behavior logic to different file/folder
